pyproject.toml
[project]
name = "slider-recognize"
version = "0.1.0"
description = "滑块验证码距离计算HTTP服务"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddddocr>=1.5.6",
    "fastapi>=0.115.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "python-multipart>=0.0.17",
    "requests>=2.32.5",
    "uvicorn>=0.32.0",
]
